Restore V2 and CCDA into separate controller
