+++ leap-ces-ccda-orchestration/src/main/resources/application-docker.yaml
