+++ leap-ces-ccda-orchestration/src/main/resources/application.yaml
@@ -1,0 +1,19 @@
+server:
+  port: 9093
+
+spring:
+    profiles:
+        active: local
+        
+logging:
+  file: /tmp/f.log
+  level:
+    root: INFO
+    gov.hhs.onc: DEBUG
+    
+nlp:
+    enabled: false
+
+info:
+  app:
+    name: LEAP CCDA Demo Orchestration
