--- leap-ces-orchestration/docker-push.sh
@@ -1,3 +1,0 @@
-docker login docker.io
-docker tag 8e09db6fa2fc ddecouteau/leap-ces-orchestration
-docker push ddecouteau/leap-ces-orchestration
