--- leap-ces-orchestration/src/main/resources/application.yaml
@@ -1,19 +1,0 @@
-server:
-  port: 9092
-
-spring:
-    profiles:
-        active: local
-        
-logging:
-  file: /tmp/f.log
-  level:
-    root: INFO
-    gov.hhs.onc: DEBUG
-    
-nlp:
-    enabled: false
-
-info:
-  app:
-    name: LEAP Demo Orchestration
