+++ leap-ces-v2-orchestration/Dockerfile
@@ -1,0 +1,5 @@
+FROM openjdk:11-jdk
+
+ARG JAR_FILE=target/leap-ces-v2-orchestration-0.0.1.war
+COPY ${JAR_FILE} app.jar
+ENTRYPOINT ["java","-Dspring.profiles.active=docker","-jar","/app.jar"]
