+++ leap-ces-v2-orchestration/docker-push.sh
@@ -1,0 +1,3 @@
+docker login docker.io
+docker tag 8e09db6fa2fc ddecouteau/leap-ces-v2-orchestration
+docker push ddecouteau/leap-ces-v2-orchestration
