+++ leap-ces-v2-orchestration/src/main/java/gov/hhs/onc/leap/ces/v2/orchestration/config/RestTemplateConfig.java
@@ -1,0 +1,14 @@
+package gov.hhs.onc.leap.ces.v2.orchestration.config;
+
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class RestTemplateConfig {
+    @Bean
+    public RestTemplate restTemplate(RestTemplateBuilder builder) {
+        return builder.build();
+    }
+}
