+++ leap-ces-v2-orchestration/src/main/resources/application-local.yaml
@@ -1,0 +1,11 @@
+cds:
+    host:
+        url: ${CDS_HOST_URL}
+    
+sls:
+    host:
+        url: ${SLS_HOST_URL}
+        
+fhir:
+  r4:
+    baseurl: ${HAPI_FHIR_URL}
